requirements.txt
deep-translator==1.11.4
discord.py==2.4.0
openai-whisper==20240930
python-dotenv==1.0.1
yt-dlp==2025.1.26
